♻️ refacto code
